# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZenFM API
  version: 0.1.0
  description: |
    The local, single-owner ZenFM HTTP API. Browser requests authenticate with
    an opaque `zenfm_session` cookie and send the session-bound value returned
    as `csrfToken` in `X-ZenFM-CSRF` for every mutation. Personal API tokens use
    `Authorization: Bearer`; their restricted scope excludes owner security,
    token management, update, and lifecycle operations.
servers:
  - url: /
paths:
...
  /api/v1/session:
    get:
      operationId: validateSession
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The validated browser session. Validation does not extend its absolute lifetime.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded and a secure session cookie was set.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly, SameSite=Strict session cookie; Secure under HTTPS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UniformLoginFailure'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: logout
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      responses:
        '204':
          description: The current session was revoked and its cookie cleared.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/owner/password:
    put:
      operationId: changeOwnerPassword
      description: Changes the sole owner's password and revokes every session and API token.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
      responses:
        '200':
          description: Password changed; a replacement browser session was issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tokens:
    get:
      operationId: listPersonalTokens
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Token metadata. Token secrets are never returned here.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonalToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createPersonalToken
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonalToken'
      responses:
        '201':
          description: The plaintext token is returned exactly once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PersonalToken'
                  - type: object
                    required: [token]
                    properties:
                      token:
                        type: string
                        writeOnly: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tokens/{tokenId}:
    delete:
      operationId: revokePersonalToken
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
        - $ref: '#/components/parameters/TokenId'
      responses:
        '204':
          description: Token revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/settings:
    get:
      operationId: getSettings
      security:
        - sessionCookie: []
        - personalToken: []
      responses:
        '200':
          description: Owner preferences and redacted effective server settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updatePreferences
      description: Updates owner presentation preferences. Personal tokens cannot change settings.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceUpdate'
      responses:
        '200':
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/files:
    get:
      operationId: listFiles
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - $ref: '#/components/parameters/Hidden'
      responses:
        '200':
          description: One directory. Special entries are metadata-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
        - $ref: '#/components/parameters/Path'
        - in: query
          name: recursive
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Target deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/directory:
    post:
      operationId: createDirectory
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [path]
              properties:
                path:
                  $ref: '#/components/schemas/Path'
      responses:
        '201':
          description: Directory created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/files/content:
    get:
      operationId: getFileContent
      description: Returns exact, bounded UTF-8 source for the text editor; unlike preview output, HTML is not transformed.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
      responses:
        '200':
          description: Exact editable text, limited to 4 MiB.
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/Unsupported'
    put:
      operationId: putFile
      description: Atomically creates or explicitly replaces a regular file. Symlinks and special files are rejected.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
        - $ref: '#/components/parameters/Path'
        - in: header
          name: If-None-Match
          schema:
            type: string
          description: Use `*` to reject overwrite conflicts.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: File created.
        '204':
          description: File replaced.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
  /api/v1/files/move:
    post:
      operationId: moveFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '204':
          description: Target moved or renamed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/files/copy:
    post:
      operationId: copyFile
      description: Recursion skips symlinks, pseudo-filesystems, and non-regular entries.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '204':
          description: Target copied.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/files/raw:
    get:
      operationId: readFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
      responses:
        '200':
          description: Streamed regular-file content with range support.
          headers:
            Accept-Ranges:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/preview:
    get:
      operationId: previewFile
      description: Returns a bounded, safe preview representation for a regular file.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - in: query
          name: width
          schema:
            type: integer
            minimum: 16
            maximum: 4096
        - in: query
          name: height
          schema:
            type: integer
            minimum: 16
            maximum: 4096
      responses:
        '200':
          description: Preview stream.
          content:
            image/png:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
            text/html:
              schema:
                type: string
              description: Sanitized HTML only; scripts, event handlers, and external resources are removed.
            text/csv:
              schema:
                type: string
            text/markdown:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
            audio/*:
              schema:
                type: string
                format: binary
            video/*:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/Unsupported'
  /api/v1/files/checksum:
    get:
      operationId: checksumFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - in: query
          name: algorithm
          schema:
            type: string
            enum: [sha256, sha512]
            default: sha256
      responses:
        '200':
          description: Digest of the regular file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checksum'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/archive:
    post:
      operationId: downloadArchive
      description: Streams an archive whose entry names are canonical and traversal-safe.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveRequest'
      responses:
        '200':
          description: Archive stream.
          content:
            application/zip:
              schema:
                type: string
                format: binary
            application/x-tar:
              schema:
                type: string
                format: binary
            application/gzip:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/usage:
    get:
      operationId: diskUsage
      security:
        - sessionCookie: []
        - personalToken: []
      responses:
        '200':
          description: Filesystem capacity for the configured root.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usage'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zenfm_session
    personalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque ZenFM personal token
...
  responses:
...
    Unauthorized:
      description: Missing, expired, or revoked credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 128
        password:
          type: string
          maxLength: 1024
...
    Session:
      type: object
      additionalProperties: false
      required: [authenticated, setupRequired]
      properties:
        authenticated:
          type: boolean
        username:
          type: string
          const: koreader
        setupRequired:
          type: boolean
        csrfToken:
          type: string
          description: Kept only in browser memory by the official client.
        idleExpiresAt:
          type: string
          format: date-time
        absoluteExpiresAt:
          type: string
          format: date-time
...